tests/pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = "smaug tests"
authors = ["Chris Guida <chrisguida@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
black = "^22.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-timeout = "^2.1.0"
# pyln-testing = "^0.12.1"
pyln-testing = "23.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
